+++ js/map.js
@@ -1,41 +1,33 @@
-//Google maps kartta ja reittietsintä nappula
-//Google API suorittaa initMap() funktion
-
+//Leaflet ja Mapbox luovat kartan
 const options = {       //Asetukset sijainnille
     enableHighAccuracy: true,
     timeout: 5000,
     maximumAge: 0
 }
 
-function initMap() {
-    const shop = { lat: 60.223886, lng: 24.758337};     //Tallennetaan haluttu karttakohta
 
-    const map = new google.maps.Map(document.getElementById("map"), {   //Luodaan uusi kartta olio
-        zoom: 11,       //Zoom taso. 1 = Koko maapallo
-        center: shop,   //Kartta keskitetään haluttuun pisteeseen
-    }, options);
+const map = L.map('map').setView([60.223886, 24.758337], 12);   //Luodaan kartta ja asetetaan se haluttuun pisteeseen
 
-    new google.maps.Marker({        //Luodaan uusi karttamerkki
-        position: shop,             //Asetetaan se haluttuun kohtaan
-        map: map,                   //Halutulle kartalle
-        title: "Mondo's Modules",   //Otsikko
-    });
+L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {     //Asetetaan kartalle pohja
+    attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
+}).addTo(map);
 
-    fetch("http://api.open-notify.org/iss-now.json")
-        .then(r => r.json())
-        .then(jsondata => {
-            let isspos = { lat: Number(jsondata.iss_position.latitude), lng: Number(jsondata.iss_position.longitude)};
-            new google.maps.Marker({
-                position: isspos,
-                map: map,
-                title: "ISS",
-            });
-    });
-}
+L.marker([60.223886, 24.758337]).addTo(map)     //Asetetaan merkki, annetaan sille nimi ja avataan se
+    .bindPopup("Mondo's Modules")
+    .openPopup();
+
+fetch("http://api.open-notify.org/iss-now.json")    //Haetaan API:sta ISS:n sijainti. Sijaintiin laitetaan merkki
+.then(r => r.json())
+.then(jsondata => {
+    L.marker([Number(jsondata.iss_position.latitude), Number(jsondata.iss_position.longitude)]).addTo(map)
+        .bindPopup("ISS-position")
+});
 
 
 const button = document.querySelector('#location');
 
+//Ei ehditty toteuttamaan Mapboxin directions ominaisuutta
+//Hakee nykyisen sijainnin ja avaa google mapsin
 button.addEventListener('click', function() {
     navigator.geolocation.getCurrentPosition(function(pos){
         console.log(pos);